+++ pioreactor/actions/leader/experiment_profile.py
@@ -613,7 +613,7 @@ def start_job(
                             },
                             "args": args,
                             "config_overrides": [
-                                f"{parent_job.job_key}.config,{key},{value}"
+                                f"{job_name}.config,{key},{value}"
                                 for (key, value) in config_overrides.items()
                             ],
                         },